[UPDATE] Se ha actualizado el orden de insertar datos, además, se ha actualizado algunos tipos de datos de la tabla medicion

+++ Practica 1/dashboard/DDL.sql
@@ -5,9 +5,9 @@ CREATE TABLE medicion (
 	id INT PRIMARY KEY AUTO_INCREMENT,
     temperatura_externa DECIMAL(10, 2),
     humedad_relativa DECIMAL(10, 2),
-    humedad_absoluta DECIMAL(10, 2),
+    humedad_absoluta DECIMAL(10, 6),
     velocidad_viento DECIMAL(10, 2),
-    direccion_viento DECIMAL(10, 2),
+    direccion_viento INT,
     presion_barometrica DECIMAL(10, 2),
     fecha_registro DATETIME
 );
